+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## [Unreleased]
 ### Added
+- OAuth 2.0 Password Grant authentication, by @HardNorth
+### Changed
+- Client version updated to [5.6.6](https://github.com/reportportal/client-Python/releases/tag/5.6.6), by @HardNorth
+### Fixed
+- Some configuration parameter names, which are different in the client, by @HardNorth
+### Removed
+- `token` param support, as it was deprecated pretty while ago, by @HardNorth
+
+## [5.0.0]
+### Added
 - Support for `Python 3.13`, by @HardNorth
 ### Changed
 - Client version updated to [5.6.5](https://github.com/reportportal/client-Python/releases/tag/5.6.5), by @HardNorth